requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
python-dateutil==2.9.0.post0